Sky illuminance values for the sun and moon

Function returns sky illuminance parameters for both the sun and the moon, in addition to some ancillary parameters such as sun and moon azimuth and altitude.

skylight(.data, longitude, latitude, date, sky_condition = 1)

Arguments

.data

A data frame or data frame extension (e.g. a tibble) with named columns: longitude, latitude, date and optionally sky_condition

longitude

decimal longitude (single value or vector of values)

latitude

decimal latitude (single value or vector of values)

date

date and time in POSIXct format with GMT/UTC as time zone (single value or vector of values)

sky_condition

a positive value (>=1) with which to scale illuminance values (1 = cloud cover < 30 3 = average clouds, 10 = dark stratus clouds). The value is set 1 (one) by default if missing (when all arguments are provided or should the column not be present in a data frame).

Value

Sun and moon illuminance values (in lux), as well as their respective location in the sky (altitude, azimuth).

Details

The code is almost verbatim transcription of the work "Computer Programs for Sun and Moon Illuminance With Contingent Tables and Diagrams by Janiczek and DeYoung" and published in the US Naval observatory circular nr. 171, 1987.

Required parameters are a location (in longitude, latitude), and a date in POSIXct format set to the GMT/UTC time zone. Conversions to GMT/UTC should be done externally, errors are not trapped.

The original code has been vectorized, as such vectors of location, time and/or sky conditions can be provided.

Examples

 # run the function on standard
 # input variables (single values or vectors of equal size)
 df <- skylight(
  longitude = -135.8,
  latitude = -23.4,
  date = as.POSIXct("1986-12-18 21:00:00", tz = "GMT"),
  sky_condition = 1
 )

 print(df)
#>   sun_azimuth sun_altitude sun_illuminance moon_azimuth moon_altitude
#> 1         346      89.9565        123786.3          282     -61.99004
#>   moon_illuminance moon_fraction total_illuminance
#> 1     1.066142e-27      93.94848          123786.3

 # create data frame of input variables
 input <- data.frame(
   longitude = 0,
   latitude = 50,
   date =  as.POSIXct("2020-06-18 00:00:00", tz = "GMT") + seq(0, 1*24*3600, 1800),
   sky_condition = 1
  )

  # calculate on data frame
  df <- skylight(input)

  print(df)
#>    longitude latitude                date sky_condition sun_azimuth
#> 1          0       50 2020-06-18 00:00:00             1         360
#> 2          0       50 2020-06-18 00:30:00             1           7
#> 3          0       50 2020-06-18 01:00:00             1          14
#> 4          0       50 2020-06-18 01:30:00             1          21
#> 5          0       50 2020-06-18 02:00:00             1          28
#> 6          0       50 2020-06-18 02:30:00             1          34
#> 7          0       50 2020-06-18 03:00:00             1          41
#> 8          0       50 2020-06-18 03:30:00             1          47
#> 9          0       50 2020-06-18 04:00:00             1          52
#> 10         0       50 2020-06-18 04:30:00             1          58
#> 11         0       50 2020-06-18 05:00:00             1          64
#> 12         0       50 2020-06-18 05:30:00             1          69
#> 13         0       50 2020-06-18 06:00:00             1          74
#> 14         0       50 2020-06-18 06:30:00             1          80
#> 15         0       50 2020-06-18 07:00:00             1          85
#> 16         0       50 2020-06-18 07:30:00             1          91
#> 17         0       50 2020-06-18 08:00:00             1          97
#> 18         0       50 2020-06-18 08:30:00             1         103
#> 19         0       50 2020-06-18 09:00:00             1         110
#> 20         0       50 2020-06-18 09:30:00             1         118
#> 21         0       50 2020-06-18 10:00:00             1         127
#> 22         0       50 2020-06-18 10:30:00             1         138
#> 23         0       50 2020-06-18 11:00:00             1         150
#> 24         0       50 2020-06-18 11:30:00             1         164
#> 25         0       50 2020-06-18 12:00:00             1         179
#> 26         0       50 2020-06-18 12:30:00             1         195
#> 27         0       50 2020-06-18 13:00:00             1         209
#> 28         0       50 2020-06-18 13:30:00             1         221
#> 29         0       50 2020-06-18 14:00:00             1         232
#> 30         0       50 2020-06-18 14:30:00             1         241
#> 31         0       50 2020-06-18 15:00:00             1         249
#> 32         0       50 2020-06-18 15:30:00             1         256
#> 33         0       50 2020-06-18 16:00:00             1         263
#> 34         0       50 2020-06-18 16:30:00             1         269
#> 35         0       50 2020-06-18 17:00:00             1         274
#> 36         0       50 2020-06-18 17:30:00             1         280
#> 37         0       50 2020-06-18 18:00:00             1         285
#> 38         0       50 2020-06-18 18:30:00             1         291
#> 39         0       50 2020-06-18 19:00:00             1         296
#> 40         0       50 2020-06-18 19:30:00             1         301
#> 41         0       50 2020-06-18 20:00:00             1         307
#> 42         0       50 2020-06-18 20:30:00             1         313
#> 43         0       50 2020-06-18 21:00:00             1         319
#> 44         0       50 2020-06-18 21:30:00             1         325
#> 45         0       50 2020-06-18 22:00:00             1         332
#> 46         0       50 2020-06-18 22:30:00             1         338
#> 47         0       50 2020-06-18 23:00:00             1         345
#> 48         0       50 2020-06-18 23:30:00             1         352
#> 49         0       50 2020-06-19 00:00:00             1         360
#>    sun_altitude sun_illuminance moon_azimuth moon_altitude moon_illuminance
#> 1   -16.6276922    2.872734e-05           40    -18.193474     7.503453e-14
#> 2   -16.3499406    3.874383e-05           46    -14.953394     2.407014e-12
#> 3   -15.4764151    9.949008e-05           52    -11.355126     1.189855e-10
#> 4   -14.0292824    4.779034e-04           58     -7.453916     8.248541e-09
#> 5   -12.0432004    4.163346e-03           64     -3.301080     6.409745e-07
#> 6    -9.5619226    6.272634e-02           69      1.419424     1.859100e-05
#> 7    -6.6346679    1.513215e+00           75      5.721701     6.080463e-05
#> 8    -3.3128799    4.883364e+01           80     10.298522     1.419753e-04
#> 9     0.7953068    1.135681e+03           86     14.988517     2.442273e-04
#> 10    4.4912107    3.588667e+03           91     19.729673     3.528487e-04
#> 11    8.6210695    8.836026e+03           97     24.473187     4.600399e-04
#> 12   12.9978263    1.647697e+04          103     29.169588     5.611704e-04
#> 13   17.5515623    2.546993e+04          109     33.763197     6.531394e-04
#> 14   22.2338494    3.514985e+04          116     38.187614     7.337182e-04
#> 15   27.0013999    4.508262e+04          123     42.360937     8.012695e-04
#> 16   31.8108170    5.494771e+04          132     46.180854     8.546214e-04
#> 17   36.6149501    6.448485e+04          141     49.521126     8.930063e-04
#> 18   41.3587950    7.347061e+04          151     52.233116     9.160327e-04
#> 19   45.9739704    8.170807e+04          163     54.158256     9.236741e-04
#> 20   50.3709721    8.902247e+04          175     55.156230     9.162685e-04
#> 21   54.4287710    9.525949e+04          187     55.143903     8.945187e-04
#> 22   57.9831218    1.002850e+05          199     54.124953     8.594760e-04
#> 23   60.8204418    1.039860e+05          211     52.188822     8.124957e-04
#> 24   62.6945544    1.062732e+05          221     49.479767     7.551679e-04
#> 25   63.3878939    1.070869e+05          230     46.158060     6.892447e-04
#> 26   62.8054871    1.064046e+05          239     42.372696     6.165788e-04
#> 27   61.0277194    1.042452e+05          246     38.249759     5.390784e-04
#> 28   58.2652189    1.006656e+05          253     33.891392     4.586801e-04
#> 29   54.7648487    9.575262e+04          259     29.379517     3.773418e-04
#> 30   50.7442529    8.961748e+04          265     24.780795     2.970683e-04
#> 31   46.3719517    8.239251e+04          271     20.151350     2.199999e-04
#> 32   41.7723192    7.423023e+04          277     15.541076     1.486377e-04
#> 33   37.0371816    6.530360e+04          282     10.998449     8.639239e-05
#> 34   32.2364330    5.580745e+04          287      6.580348     3.877221e-05
#> 35   27.4260094    4.596257e+04          293      2.399543     1.282259e-05
#> 36   22.6535762    3.602524e+04          299     -2.023007     1.024176e-06
#> 37   17.9627190    2.630907e+04          304     -5.941592     1.913078e-08
#> 38   13.3965732    1.723440e+04          310     -9.586631     3.561675e-10
#> 39    9.0027551    9.438297e+03          316    -12.908029     9.178562e-12
#> 40    4.8476623    3.931513e+03          323    -15.853258     3.642676e-13
#> 41    1.0980453    1.291022e+03          330    -18.368778     2.375493e-14
#> 42   -2.9952932    6.670322e+01          337    -20.402471     2.668929e-15
#> 43   -6.3482060    2.059679e+00          344    -21.907030     5.350966e-16
#> 44   -9.3110728    8.250585e-02          351    -22.843968     1.962142e-16
#> 45  -11.8324066    5.241369e-03          359    -23.187560     1.335644e-16
#> 46  -13.8626994    5.728062e-04            6    -22.927820     1.697192e-16
#> 47  -15.3576065    1.131358e-04           13    -22.071709     4.011410e-16
#> 48  -16.2815607    4.170703e-05           21    -20.642229     1.741070e-15
#> 49  -16.6112142    2.924137e-05           28    -18.675657     1.356837e-14
#>    moon_fraction total_illuminance
#> 1      10.648745      5.287273e-04
#> 2      10.522756      5.387438e-04
#> 3      10.397380      5.994902e-04
#> 4      10.272619      9.779116e-04
#> 5      10.148477      4.663987e-03
#> 6      10.024955      6.324493e-02
#> 7       9.902057      1.513775e+00
#> 8       9.779786      4.883428e+01
#> 9       9.658143      1.135681e+03
#> 10      9.537132      3.588667e+03
#> 11      9.416756      8.836027e+03
#> 12      9.297017      1.647697e+04
#> 13      9.177918      2.546993e+04
#> 14      9.059461      3.514985e+04
#> 15      8.941649      4.508262e+04
#> 16      8.824486      5.494771e+04
#> 17      8.707973      6.448485e+04
#> 18      8.592114      7.347061e+04
#> 19      8.476911      8.170807e+04
#> 20      8.362366      8.902247e+04
#> 21      8.248484      9.525949e+04
#> 22      8.135265      1.002850e+05
#> 23      8.022713      1.039860e+05
#> 24      7.910831      1.062732e+05
#> 25      7.799621      1.070869e+05
#> 26      7.689086      1.064046e+05
#> 27      7.579229      1.042452e+05
#> 28      7.470052      1.006656e+05
#> 29      7.361558      9.575262e+04
#> 30      7.253750      8.961749e+04
#> 31      7.146630      8.239251e+04
#> 32      7.040201      7.423023e+04
#> 33      6.934466      6.530360e+04
#> 34      6.829427      5.580745e+04
#> 35      6.725088      4.596257e+04
#> 36      6.621449      3.602525e+04
#> 37      6.518516      2.630907e+04
#> 38      6.416289      1.723440e+04
#> 39      6.314772      9.438297e+03
#> 40      6.213967      3.931513e+03
#> 41      6.113877      1.291022e+03
#> 42      6.014504      6.670372e+01
#> 43      5.915852      2.060179e+00
#> 44      5.817922      8.300585e-02
#> 45      5.720718      5.741369e-03
#> 46      5.624242      1.072806e-03
#> 47      5.528496      6.131358e-04
#> 48      5.433484      5.417070e-04
#> 49      5.339207      5.292414e-04

  # the above statement can also be used
  # in a piped fashion in R >= 4.2
  # input |> skylight()